+++ nHallRandom.h
@@ -9,4 +9,4 @@
 enum NH_Cases {N200,N100,N50,P25,P50,P100,P200};
 std::vector<NH_Cases> NH_DECK{};
 
-NH_Cases NHall_Card_Generator(float seed);
+NH_Cases NHall_Card_Generator(unsigned seed);
