+++ slcRandom.h
@@ -9,6 +9,6 @@
 enum SLC_Cases {B3,B2,B1,F1,F2,F3,DC,OSAP};
 std::vector<SLC_Cases> SLC_DECK{};
 
-SLC_Cases SLC_Card_Generator(float seed);
+SLC_Cases SLC_Card_Generator(unsigned seed);
